backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
beautifulsoup4==4.13.5
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.2
django-rest-framework==0.1.0
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
kombu==5.5.4
lxml==6.0.2
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==6.4.0
requests==2.32.5
rest-framework-simplejwt==0.0.2
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
